tests/wpt/web-platform-tests/css/css-animations/parsing/animation-direction-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Animations: parsing animation-direction with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-animations/#propdef-animation-direction">
<meta name="assert" content="animation-direction supports only the grammar '<single-animation-direction> #'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("animation-direction", "auto");
test_invalid_value("animation-direction", "normal reverse");
</script>
</body>
</html>
